+++ src/main/java/com/sparta/deliverit/order/presentation/dto/response/MenuInfo.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class MenuInfo {
     @JsonProperty("menuName")
@@ -13,16 +15,16 @@ public class MenuInfo {
     private final int quantity;
 
     @JsonProperty("price")
-    private final int price;
+    private final BigDecimal price;
 
     @Builder
-    private MenuInfo(String menuName, int quantity, int price) {
+    private MenuInfo(String menuName, int quantity, BigDecimal price) {
         this.menuName = menuName;
         this.quantity = quantity;
         this.price = price;
     }
 
-    public static MenuInfo create(String menuName, int quantity, int price) {
+    public static MenuInfo create(String menuName, int quantity, BigDecimal price) {
         return MenuInfo.builder()
                 .menuName(menuName)
                 .quantity(quantity)
